borad 스펠링 수정 및 XML 총 수정
