--- app/src/main/res/layout/activity_borad_detail.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".activites.BoradDetail"
-    android:padding="16dp">
-    <RelativeLayout
-        android:id="@+id/layout_title"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-        <de.hdodenhof.circleimageview.CircleImageView
-            android:id="@+id/iv_main"
-            android:layout_width="30dp"
-            android:layout_height="30dp"
-            android:src="@drawable/profle"/>
-        <TextView
-            android:id="@+id/nickname"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="닉네임"
-            android:layout_toRightOf="@id/iv_main"/>
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text=","
-            android:layout_toRightOf="@id/nickname"
-            android:layout_toLeftOf="@id/age"/>
-        <TextView
-            android:id="@+id/age"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="나이"
-            android:layout_toRightOf="@id/nickname"/>
-    </RelativeLayout>
-    <ImageView
-        android:id="@+id/iv"
-        android:layout_width="match_parent"
-        android:layout_height="450dp"
-        android:background="#81FFFFFF"
-        android:layout_below="@id/layout_title"/>
-    <RelativeLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/iv">
-        <ImageView
-            android:id="@+id/request"
-            android:layout_width="30dp"
-            android:layout_height="53dp"
-            android:src="@drawable/baseline_request"
-            android:layout_alignParentLeft="true"
-            android:layout_marginLeft="70dp"/>
-        <ImageView
-            android:id="@+id/chat"
-            android:layout_width="30dp"
-            android:layout_height="50dp"
-            android:src="@drawable/baseline_send_24"
-            android:layout_alignParentRight="true"
-            android:layout_marginRight="70dp"/>
-    </RelativeLayout>
-
-</RelativeLayout>
